UI cleanup, API cleanup, code cleanup
Created a cleaner thunk to handle asynchronous data fetched from API using AXIOS. Removed fetchSuccess reducer and migrated its functionality inside an extra reducer used for successful completion of asynchronous request from the new thunk. Added a new mutation query for liking and unliking posts, presentation is rendered through postSlice in a newly created function called updateLike which finds the existing array and adds or removes 1 from the quantity of likes alongside sets it as liked or not which is necessary to identify whether we are liking or un-liking our post in the endpoint. This presentation is rendered for users convenience until it is able to be re-fetched from the backend.
Globally store the state of the feed the user last used. If user creates a new post, user is redirected to the timeline with the feed they just posted to. Feed is now set globally. Currently in the postSlice for now, but as the project expands... the plan is to separate of concerns.